Added an extension method to easily add requesthandlers as services Changed the ConfigureServices method to use this extension method. Removed the RequestHandlerResolver because I added it as a default resolver for the Mvc package. This could be changed in a later update, because this way we force the user to use the built-in IoC-container.
There were no breaking changes in this update.
